lab/load-test/requirements.txt
locust
python-dotenv